+++ src/trunk/apps/tools/scconfig/panels/inventory.h
@@ -105,6 +105,7 @@ class InventoryPanel : public ConfiguratorPanel {
 		void inspectFile();
 
 		void import();
+		void testInventory();
 		void testSync();
 		void sync();
 		void syncKeys();